Clang Warnings, External Linking, Type and Lifecycle Hardening
General:
* updated order of compile flags to force several clang inspections off
* separated internal and external includes and linking for cmake targets

Assets:
* changed the internal per asset storage object to Arci<...>
* replaced exposed asset types to shared lifetime control over internal resources
* replaced several usages of raw pointers with references to harden the usage against nullptr ( instead of excessive checks )
* corrected several jsdoc annotations
* added a new helper wrapper to consume asset_guid from an already instantiated asset as a source
** by default it will constrain the source to align with the requested instantiation type, otherwise reject the asset_guid assignment
** the TypedAssetGuid can be copy/move assigned and constructed as long the inheritance reflects an upward cast in strict mode

Core:
* changed several wrappers from raw pointers to non-owning wrappers or ownership-sharing alternatives

Common:
* implemented concurrent move-assignments for Arci<...> objects
* corrected move-constructions based on the new schema for concurrent move assignments of Arci<...> objects
* added support for incomplete Arci<...> types within interfaces
* widened casting behaviour of memory pointers to allow wildcard casting from void sources

Editor:
* disabled several paths of unsafe or incomplete asset instantiation and registration

Events:
* replaced SFINAE of event emitter constructors with concepts

Reflect:
* reworked the internal hierarchy reconstruction by introducing a deferred template instantiation
** this should prevent cyclic and immediate instantiation while consuming the derived types, resulting in errors due to incomplete types
** fixes meta class template instantiation with clang due to previously failing constraints within template specialization requesting a complete type
* enforced complete type constraints at TypedMetaClass base template

Reflow:
* changed the font attribute to just hold the asset_guid of the font to use instead of a unsecure raw pointer ( preventing several lifecycle problems )

+++ src/Engine.ACS/Traits.hpp
@@ -33,7 +33,7 @@ namespace hg {
 		static std::atomic_uint_fast32_t ai { 1 };
 		const auto pre = ai.fetch_add(1, std::memory_order_relaxed);
 		const auto post = std::chrono::high_resolution_clock::now().time_since_epoch().count();
-		return { pre, 0, 0, post };
+		return { pre, 0u, 0u, static_cast<u64>(post) };
 	}
 }
